+++ src/snippets/fullApp/activeTabAnswers.ts
@@ -27,9 +27,9 @@ embed
 	});
 // Functions to show and hide a loader while iframe loads
 function showLoader() {
-	document.getElementById('loader').style.display = 'block';
+	// document.getElementById('loader').style.display = 'block';
 }
 function hideLoader() {
-	document.getElementById('loader').style.display = 'none';
+	// document.getElementById('loader').style.display = 'none';
 }
 `;
