+++ src/snippets/fullApp/activeTagPinboards.ts
@@ -26,9 +26,9 @@ embed
 	});
 // Functions to show and hide a loader while iframe loads
 function showLoader() {
-	document.getElementById('loader').style.display = 'block';
+	// document.getElementById('loader').style.display = 'block';
 }
 function hideLoader() {
-	document.getElementById('loader').style.display = 'none';
+	// document.getElementById('loader').style.display = 'none';
 }
 `;
