+++ src/snippets/pinboard/hiddenActions.ts
@@ -38,11 +38,11 @@ try {
 
 // Functions to show and hide a loader while iframe loads
 function showLoader() {
-	document.getElementById('loader').style.display = 'block';
+	// document.getElementById('loader').style.display = 'block';
 }
 
 function hideLoader() {
-	document.getElementById('loader').style.display = 'none';
+	// document.getElementById('loader').style.display = 'none';
 }
 
 // Functions to show or hide No data images
