+++ src/snippets/pinboard/modifyAvailableAction.ts
@@ -40,11 +40,11 @@ try {
 
 // Functions to show and hide a loader while iframe loads
 function showLoader() {
-	document.getElementById('loader').style.display = 'block';
+	// document.getElementById('loader').style.display = 'block';
 }
 
 function hideLoader() {
-	document.getElementById('loader').style.display = 'none';
+	// document.getElementById('loader').style.display = 'none';
 }
 
 // Functions to show or hide No data images
