+++ src/snippets/search/modifyAvailableAction.ts
@@ -39,9 +39,9 @@ tsSearch
 
 // Show/hide a loader while iframe is loading
 function showLoader() {
-	document.getElementById('loader').style.display = 'block';
+	// document.getElementById('loader').style.display = 'block';
 }
 function hideLoader() {
-	document.getElementById('loader').style.display = 'none';
+	// document.getElementById('loader').style.display = 'none';
 }
 `;
