+++ src/snippets/visualization/disableActionWithComment.ts
@@ -42,10 +42,10 @@ try {
 }
 // Functions to show/hide a loader while the iframe loads
 function showLoader() {
-	document.getElementById('loader').style.display = 'block';
+	// document.getElementById('loader').style.display = 'block';
 }
 function hideLoader() {
-	document.getElementById('loader').style.display = 'none';
+	// document.getElementById('loader').style.display = 'none';
 }
 // Functions to show or hide No data images
 function showNoDataImage() {
