+++ src/typings.d.ts
@@ -1,0 +1,4 @@
+declare module 'safe-eval' {
+    const content: string;
+    export default content;
+}
